Refactored a lot. Django has been replaced by flask.
After discovering that Django reaally excpects you to use one of their database engine backend as
part of the application, I realized it was probably too much for our needs.

Flask is a much simpler framework. We don't need authentication, we have limited pages and
resources, and our data processing will happen using a different API. Flask is easily extensible
and we have a basic server up and running.

The front page (the actual map) is currently blank. However any unkown proxy used will be redirected to a
404 page which is full operational! Code was sourced from codepen and slighty modified to suit our application.