Added index.html file for base map page. This was taken from the following git resource:
https://github.com/ryanj/flask-postGIS/blob/master/templates/index.html

This file will probably be modified to suit our needs, but is exactly what we want for now.

Future work will include adding a new backend source of data that we can overlay onto this map.

Likely the national parks will be removed, but we may keep it as a side feature ;)