Added some more functionality to data module and corrsponding tests to test module.
This update includes the basis for working with a cKDTree, the c based kd-tree data
 structure from scipy.spatial library.

Also added an argument parser to specify whether or not verbose output should be requested.

Everything works so far here.

Included some edception handlers for imports just in case we want to test some things
on a platform that does not include these scientific libraries, but is still capable of
running some of the basic functionality in the project.