OK the data module is replaced by kd. It will be removed in the next push as part of depractaion.
The kd module is essentially performing the same thing, except using Solr now instead of EDGE.
 This comes with numerous advantages, at the cost of having to be run the on the server where it is at.

The major changes for this update include populating a kd tree before launching the web service.

All data points from samos are put into numpy arrays in memory, and the points are loaded into a
scipy.spatial.cKDTree. When the front end sends the map bounds to the server, the tree is queried
for nearest neighbors to two rough center points on the map (left and right). This is for better sampling.

The tree returns indices into the orignal data arrays, so thus we can have access to ancillary data via
parallel arrays. This is particulary useful with the meta strings - essentially a unique record identifier
of sorts. When a pin is created on the map initally, its content is set with the index from the server.
When clicking on a marker, we query the server with the idx so it can retrieve ancillary data for the user.
This data should be placed on the marker popup on the screen.

Currently the query is sent to the server fine, but I havent figured out how to get it to show up on the marker yet.
It still displays the original content (the idx).

Currently testing using only a very small subset of samos data. We will have to test various sizes of data that will
work for the server.